[#173] refactor: 게시글신고, 댓글신고 레이어드 아키텍처 적용
