+++ src/main/java/com/runninghi/runninghibackv2/domain/repository/ReplyReportRepository.java
@@ -1,0 +1,17 @@
+package com.runninghi.runninghibackv2.domain.repository;
+
+import com.runninghi.runninghibackv2.domain.entity.Reply;
+import com.runninghi.runninghibackv2.domain.entity.ReplyReport;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ReplyReportRepository extends JpaRepository<ReplyReport, Long> {
+
+    Page<ReplyReport> findAllByReportedReply(Reply reply, Pageable pageable);
+    List<ReplyReport> findAllByReportedReply(Reply reply);
+}
