--- src/main/java/com/runninghi/runninghibackv2/postreport/domain/aggregate/vo/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.runninghibackv2.postreport.domain.aggregate.vo;
-
-public class test {
-}
