--- src/main/java/com/runninghi/runninghibackv2/replyreport/infrastructure/repository/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.runninghibackv2.replyreport.infrastructure.repository;
-
-public class test {
-}
